remove rusqlite from main errors module
