+++ crates/rust-client/src/store/sqlite_store/notes.rs
@@ -273,7 +273,7 @@ impl SqliteStore {
             .query_map([unspent_filters], |row| row.get(0))
             .expect("no binding parameters used in query")
             .map(|result| {
-                result.map_err(|err| StoreError::ParsingError(err.to_string())).and_then(
+                result.map_err(|err| StoreError::ParsingError(err.to_string()).into()).and_then(
                     |v: String| {
                         Digest::try_from(v).map(Nullifier::from).map_err(StoreError::HexParseError)
                     },
@@ -362,9 +362,7 @@ pub(super) fn insert_input_note_tx(
             ":ignored": ignored,
             ":imported_tag": imported_tag,
         },
-    )
-    .map_err(|err| StoreError::QueryError(err.to_string()))
-    .map(|_| ())?;
+    )?;
 
     const QUERY: &str =
         "INSERT OR REPLACE INTO notes_scripts (script_hash, serialized_note_script) VALUES (?, ?)";
@@ -403,16 +401,12 @@ pub fn insert_output_note_tx(
             ":ignored": false,
             ":imported_tag": None::<u32>,
         },
-    )
-    .map_err(|err| StoreError::QueryError(err.to_string()))
-    .map(|_| ())?;
+    )?;
 
     if note_script_hash.is_some() {
         const QUERY: &str =
             "INSERT OR REPLACE INTO notes_scripts (script_hash, serialized_note_script) VALUES (?, ?)";
-        tx.execute(QUERY, params![note_script_hash, serialized_note_script,])
-            .map_err(|err| StoreError::QueryError(err.to_string()))
-            .map(|_| ())?;
+        tx.execute(QUERY, params![note_script_hash, serialized_note_script,])?;
     }
 
     Ok(())
@@ -433,8 +427,7 @@ pub fn update_note_consumer_tx_id(
             ":submitted_at": Utc::now().timestamp(),
             ":status": NOTE_STATUS_PROCESSING,
         },
-    )
-    .map_err(|err| StoreError::QueryError(err.to_string()))?;
+    )?;
 
     const UPDATE_OUTPUT_NOTES_QUERY: &str = "UPDATE output_notes SET status = :status, consumer_transaction_id = :consumer_transaction_id, submitted_at = :submitted_at WHERE note_id = :note_id;";
 
@@ -446,8 +439,7 @@ pub fn update_note_consumer_tx_id(
             ":submitted_at": Utc::now().timestamp(),
             ":status": NOTE_STATUS_PROCESSING,
         },
-    )
-    .map_err(|err| StoreError::QueryError(err.to_string()))?;
+    )?;
 
     Ok(())
 }
